♻️ refactor: 중복되는 tsx 컴포넌트로 분리 (Jira Ticket : CCC-49)
