+++ components/common/lottie/LottieAnimation.tsx
